maximum matching in the bipartite graph - passed
I think there is a better way to utilize adjacency matrix. The reason I
did not use adjacency matrix is that I was stuck at running
shortest-path algorithm on it and manipulating it.

I transform it into a simple graph represented by dictionary. The reason
we don't need a full-fledged graph as in ex1 is because this flow has
constraints, for example, the capacity, if existed, is 1.

Notice that this graph is used as residual network. When I add a
flow to an edge, I remove that edge and add a reversed one - removal
means that the flow is saturated, reverse edge means that now we allow a
reverse flow.

I tried both BFS (Edmonds-Karp) and DFS (Ford-Fulkerson) - DFS doesn't pass.